Change multiparamter extension to use AR type classes
Multiparamter extension used to do a lot of crappy work to ensure that
the invalid Date dates were not accepted, but also to store a
before_type_cast value in cases when it was invalid. We no longer need
to handle the before_type_cast value as ActiveRecord does that nicely.

Now we just ensure values to cast with a invalid date parts is not
accepted.